BleSensorSDK18: remove the "connectingTimeout"
This "feature" is harmful because after waiting for 60 seconds for the
BLE sensor to appear, AAT ultimately gives up and never tries again.
This means that a heart rate sensor which got disconnected for more
than 60 seconds will cease to work for the rest of the ride unless the
user manually reconnects (which he only does if he notices the
problem).

There's no point in having this timeout; an application which tells
Android to connect to a certain BLE sensor can do so indefinitely, and
whenever the sensor becomes available, it will just work.  If it never
becomes available (if one rides a different bike or didn't equip the
heart rate sensor), that's no problem.  So let's just remove this!