SensorList: eliminate add(SensorInterface)
All of those calls are redundant; instead, pass the SensorListItem
pointer (which the owning object already has) to each sensor.

This also eliminates some confusion I had when I read the code: this
method signature pretends it registers the SensorInterface pointer,
but when it fact all it does is register its address and name.  What
really registers the SensorInterface is the method
SensorListItem.lock().