tracker/AttributesCollector: disable LOG_INTERVAL
By setting this to zero, AAT logs the extensions with each GPS fix,
instead of waiting 3 or 4 seconds.

This gives analysis software a better accuracy; not only because of
the finer-grained interval: if AAT records a sensor value, it's an
arbitrary value, and not the average since the last recorded value.
This means that the value may contain more noise than the sensor
actually provided.

And it solves a problem with Strava: if a GPS fix doesn't contain a
"cadence" value, Strava just uses the previous value; but if there's
no "power" value, Strava instead assumes zero, resulting in very low
(and bogus) average power values.

Let's drop this limitation - this causes GPX files to grow, but the
90ies are gone, and today, we have gigabytes of free storage for GPX
files.